server/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b1
boto3==1.17.79
botocore==1.20.79
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.5
Django==3.2.3
django-allauth==0.44.0
django-environ==0.4.5
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
jmespath==0.10.0
jwcrypto==0.8
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
